# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageMeta": {
        "description": "Pagination metadata attached to every list response.",
        "properties": {
          "count": {
            "description": "Number of items in this page.",
            "title": "Count",
            "type": "integer"
          },
          "has_more": {
            "description": "True if another page is likely available.",
            "title": "Has More",
            "type": "boolean"
          },
          "limit": {
            "description": "Page size that was applied.",
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "description": "Number of items skipped.",
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Total items matching the filter, when cheaply countable.",
            "title": "Total"
          }
        },
        "required": [
          "limit",
          "offset",
          "count",
          "has_more"
        ],
        "title": "PageMeta",
        "type": "object"
      },
      "RulePage": {
        "description": "A page of loaded rules.",
        "properties": {
          "items": {
            "description": "Rules in this page, ordered by name.",
            "items": {
              "$ref": "#/components/schemas/RuleView"
            },
            "title": "Items",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta",
            "description": "Pagination metadata."
          }
        },
        "required": [
          "items",
          "meta"
        ],
        "title": "RulePage",
        "type": "object"
      },
...
      "RuleToggle": {
        "description": "Request body for enabling or disabling a rule.",
        "properties": {
          "enabled": {
            "description": "Desired state.",
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "RuleToggle",
        "type": "object"
      },
      "RuleView": {
        "description": "A loaded detection rule and its current state.",
        "properties": {
          "conditions": {
            "description": "Conditions, all of which must match.",
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Conditions",
            "type": "array"
          },
          "enabled": {
            "description": "Whether the rule is currently evaluated.",
            "title": "Enabled",
            "type": "boolean"
          },
          "group_by": {
            "description": "Packet field the window aggregates on.",
            "title": "Group By",
            "type": "string"
          },
          "name": {
            "description": "Unique rule name.",
            "title": "Name",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity",
            "description": "Severity raised when the rule fires."
          },
          "source_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Originating YAML file.",
            "title": "Source Path"
          },
          "threshold": {
            "description": "Matches required within the window before firing.",
            "title": "Threshold",
            "type": "integer"
          },
          "window": {
            "description": "Aggregation window in seconds; 0 means single-packet.",
            "title": "Window",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "severity",
          "enabled",
          "window",
          "threshold",
          "group_by",
          "conditions"
        ],
        "title": "RuleView",
        "type": "object"
      },
      "Severity": {
        "description": "Alert severity levels ordered from lowest to highest.",
        "enum": [
          "info",
          "low",
          "medium",
          "high",
          "critical"
        ],
        "title": "Severity",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "\nREST API for **Network Defender**, a modular Python network intrusion\ndetection system.\n\nEvery endpoint delegates to the internal SDK: no business logic lives in a\nroute handler, so the API, CLI and dashboard cannot drift apart.\n\n**Authentication** \u2014 when an `API_KEY` is configured, requests must carry an\n`X-API-Key` header. With no key configured, authentication is disabled so\nlocal development works out of the box.\n",
    "title": "Network Defender API",
    "version": "1.00"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/rules": {
      "get": {
        "description": "Return the currently loaded rule set, ordered by name.",
        "operationId": "list_rules_api_v1_rules_get",
        "parameters": [
          {
            "description": "Maximum number of items to return.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "description": "Maximum number of items to return.",
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Number of items to skip.",
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Number of items to skip.",
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RulePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List detection rules",
        "tags": [
          "rules"
        ]
      }
    },
...
    "/api/v1/rules/{name}": {
      "get": {
        "description": "Return a single rule by name.\n\nRaises:\n    NotFoundError: If no rule has this name.",
        "operationId": "get_rule_api_v1_rules__name__get",
        "parameters": [
          {
            "description": "Rule name.",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "description": "Rule name.",
              "maxLength": 128,
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get a rule",
        "tags": [
          "rules"
        ]
      },
      "patch": {
        "description": "Enable or disable a rule at runtime, without touching its YAML file.\n\nRaises:\n    NotFoundError: If no rule has this name.",
        "operationId": "toggle_rule_api_v1_rules__name__patch",
        "parameters": [
          {
            "description": "Rule name.",
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "description": "Rule name.",
              "maxLength": 128,
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleToggle"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enable or disable a rule",
        "tags": [
          "rules"
        ]
      }
    },
...
  },
...
}